server: use native http module to get status codes
Before there was a `configs` folder which the status
codes allowed were resided. However, that is not
necessary, the native `http` node module has a enum
with all the status codes allowed.

This commit removes the folder and httpStatusCodes.js
file, in favor of using the native approach so that,
even if for some reason they change (unlikely), the
responsibility will remain to the `http` module.

Here it follows the documentation for the STATUS_CODES
variable from the `http` native node module:

https://nodejs.org/api/http.html#http_http_status_codes